+++ NEWS.md
@@ -128,3 +128,4 @@
 
 -   Correction on add_scenarios and sdm_area to allow the usage of a character (automatizing the import of data through these functions).
 -   Use of Rmd in documentation.
+-   Change in add_scenarios naming.